+++ Private/Get-ChatCompletionMessage.ps1
@@ -35,17 +35,7 @@ function Get-ChatCompletionMessage {
         [int]$MaxRetryCount = 0,
 
         [Parameter()]
-        [OpenAIApiType]$ApiType = [OpenAIApiType]::OpenAI,
-
-        [Parameter()]
         [System.Uri]$ApiBase,
-
-        [Parameter(DontShow)]
-        [string]$ApiVersion,
-
-        [Parameter()]
-        [ValidateSet('openai', 'azure', 'azure_ad')]
-        [string]$AuthType = 'openai',
 
         [Parameter()]
         [securestring][SecureStringTransformation()]$ApiKey,
@@ -107,7 +97,6 @@ function Get-ChatCompletionMessage {
                     TimeoutSec        = $OpenAIParameter.TimeoutSec
                     MaxRetryCount     = $OpenAIParameter.MaxRetryCount
                     ApiKey            = $OpenAIParameter.ApiKey
-                    AuthType          = $OpenAIParameter.AuthType
                     Organization      = $OpenAIParameter.Organization
                     AdditionalQuery   = $AdditionalQuery
                     AdditionalHeaders = $AdditionalHeaders