+++ Public/Audio/Request-AudioSpeech.ps1
@@ -52,17 +52,7 @@ function Request-AudioSpeech {
         [int]$MaxRetryCount = 0,
 
         [Parameter()]
-        [OpenAIApiType]$ApiType = [OpenAIApiType]::OpenAI,
-
-        [Parameter()]
         [System.Uri]$ApiBase,
-
-        [Parameter(DontShow)]
-        [string]$ApiVersion,
-
-        [Parameter()]
-        [ValidateSet('openai', 'azure', 'azure_ad')]
-        [string]$AuthType = 'openai',
 
         [Parameter()]
         [securestring][SecureStringTransformation()]$ApiKey,
@@ -83,7 +73,7 @@ function Request-AudioSpeech {
 
     begin {
         # Get API context
-        $OpenAIParameter = Get-OpenAIAPIParameter -EndpointName 'Audio.Speech' -Parameters $PSBoundParameters -Engine $Model -ErrorAction Stop
+        $OpenAIParameter = Get-OpenAIAPIParameter -EndpointName 'Audio.Speech' -Parameters $PSBoundParameters -ErrorAction Stop
     }
 
     process {
@@ -135,7 +125,6 @@ function Request-AudioSpeech {
                 TimeoutSec        = $OpenAIParameter.TimeoutSec
                 MaxRetryCount     = $OpenAIParameter.MaxRetryCount
                 ApiKey            = $OpenAIParameter.ApiKey
-                AuthType          = $OpenAIParameter.AuthType
                 Organization      = $OpenAIParameter.Organization
                 Body              = $PostBody
                 AdditionalQuery   = $AdditionalQuery