+++ Public/Get-OpenAIContext.ps1
@@ -9,10 +9,7 @@ function Get-OpenAIContext {
 
     [PSCustomObject]@{
         ApiKey        = $Global:PSOpenAIContextDictionary['ApiKey']
-        ApiType       = $Global:PSOpenAIContextDictionary['ApiType']
         ApiBase       = $Global:PSOpenAIContextDictionary['ApiBase']
-        ApiVersion    = $Global:PSOpenAIContextDictionary['ApiVersion']
-        AuthType      = $Global:PSOpenAIContextDictionary['AuthType']
         Organization  = $Global:PSOpenAIContextDictionary['Organization']
         TimeoutSec    = $Global:PSOpenAIContextDictionary['TimeoutSec']
         MaxRetryCount = $Global:PSOpenAIContextDictionary['MaxRetryCount']