+++ Public/VectorStores/Batches/Get-VectorStoreFileBatch.ps1
@@ -28,17 +28,7 @@ function Get-VectorStoreFileBatch {
         [int]$MaxRetryCount = 0,
 
         [Parameter()]
-        [OpenAIApiType]$ApiType = [OpenAIApiType]::OpenAI,
-
-        [Parameter()]
         [System.Uri]$ApiBase,
-
-        [Parameter(DontShow)]
-        [string]$ApiVersion,
-
-        [Parameter()]
-        [ValidateSet('openai', 'azure', 'azure_ad')]
-        [string]$AuthType = 'openai',
 
         [Parameter()]
         [securestring][SecureStringTransformation()]$ApiKey,
@@ -96,7 +86,6 @@ function Get-VectorStoreFileBatch {
             TimeoutSec        = $OpenAIParameter.TimeoutSec
             MaxRetryCount     = $OpenAIParameter.MaxRetryCount
             ApiKey            = $OpenAIParameter.ApiKey
-            AuthType          = $OpenAIParameter.AuthType
             Headers           = @{'OpenAI-Beta' = 'assistants=v2' }
             Organization      = $OpenAIParameter.Organization
             AdditionalQuery   = $AdditionalQuery