+++ Public/VectorStores/Batches/Get-VectorStoreFileInBatch.ps1
@@ -49,17 +49,7 @@ function Get-VectorStoreFileInBatch {
         [int]$MaxRetryCount = 0,
 
         [Parameter()]
-        [OpenAIApiType]$ApiType = [OpenAIApiType]::OpenAI,
-
-        [Parameter()]
         [System.Uri]$ApiBase,
-
-        [Parameter(DontShow)]
-        [string]$ApiVersion,
-
-        [Parameter()]
-        [ValidateSet('openai', 'azure', 'azure_ad')]
-        [string]$AuthType = 'openai',
 
         [Parameter()]
         [securestring][SecureStringTransformation()]$ApiKey,
@@ -145,7 +135,6 @@ function Get-VectorStoreFileInBatch {
                     TimeoutSec        = $OpenAIParameter.TimeoutSec
                     MaxRetryCount     = $OpenAIParameter.MaxRetryCount
                     ApiKey            = $OpenAIParameter.ApiKey
-                    AuthType          = $OpenAIParameter.AuthType
                     Organization      = $OpenAIParameter.Organization
                     Headers           = @{'OpenAI-Beta' = 'assistants=v2' }
                     AdditionalQuery   = $AdditionalQuery